requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
fastapi.middleware.cors